Refactor cavy-cli into separate reporter
