+++ src/TestRunner.js
@@ -11,10 +11,11 @@
 //              tests.
 export default class TestRunner {
 
-  constructor(component, testSuites, startDelay, sendReport) {
+  constructor(component, testSuites, startDelay, reporter, sendReport) {
     this.component = component;
     this.testSuites = testSuites;
     this.startDelay = startDelay;
+    this.sendReport = reporter;
     // Using the sendReport prop is deprecated - cavy checks whether the
     // cavy-cli server is listening and sends a report if true.
     this.shouldSendReport = sendReport;
@@ -48,13 +49,6 @@ export default class TestRunner {
     const duration = (stop - start) / 1000;
     console.log(`Cavy test suite stopped at ${stop}, duration: ${duration} seconds.`);
 
-    // Compile the report object.
-    const report = {
-      results: this.testResults,
-      errorCount: this.errorCount,
-      duration: duration
-    }
-
     if (this.shouldSendReport != undefined) {
       const message = 'Deprecation warning: using the `sendReport` prop is ' +
                       'deprecated. By default, Cavy now checks whether the ' +
@@ -64,9 +58,16 @@ export default class TestRunner {
 
       if (!this.shouldSendReport) return;
     }
+    
+    // Compile the report object.
+    const report = {
+      results: this.testResults,
+      errorCount: this.errorCount,
+      duration: duration
+    }
 
     // Send report to reporter (default is cavy-cli)
-    await this.reportResults(report);
+    await this.sendReport(report);
   }
 
   // Internal: Synchronously runs each test case within a test suite, outputting
@@ -98,54 +99,6 @@ export default class TestRunner {
       this.testResults.push({message: errorMsg, passed: false});
       // Increase error count for reporting.
       this.errorCount += 1;
-    }
-  }
-
-  // Internal: Check that cavy-cli server is running and if so, send report.
-  async reportResults(report) {
-    const url = 'http://127.0.0.1:8082/';
-
-    try {
-      const response = await fetch(url);
-      const text = await response.text();
-
-      if (text == 'cavy-cli running') {
-        return this.send(report);
-      } else {
-        throw new Error('Unexpected response');
-      }
-    } catch (e) {
-      console.log(`Skipping sending test report to cavy-cli - ${e.message}.`)
-    }
-  }
-
-  // Internal: Make a post request to the cavy-cli server with the test report.
-  async send(report) {
-    const url = 'http://127.0.0.1:8082/report';
-
-    const options = {
-      method: 'POST',
-      body: JSON.stringify(report),
-      headers: { 'Content-Type': 'application/json' }
-    };
-
-    try {
-      await fetch(url, options);
-      console.log('Cavy test report successfully sent to cavy-cli');
-    } catch (e) {
-      this.handleError(e, url);
-    }
-  }
-
-  handleError(error, url) {
-    if (error.message.match(/Network request failed/)) {
-      console.group(`Cavy test report server is not running at ${url}`);
-      console.log("If you are using cavy-cli, maybe it's not set up correctly or not reachable from this device?");
-      console.groupEnd();
-    } else {
-      console.group('Error sending test results')
-      console.warn(error.message);
-      console.groupEnd();
     }
   }
